Kin018 sandbox (#26)
* changed paragraph in services

* same

* Added possible pictures for AboutMe

* Added Photos and initial infinite carousel completed

* Added pause on hover

* Possible fade in and fade out but still pondreing if I want it

* hover pause

* Added Slider containre

* Used https://screely.com/

to make screenshot

* Setup EmailJs Service

Indeed, someone could copy your keys (a well-known issue of any public API), but they will only be able to send your templates with your content, and they will not be able to send a custom email with their content (spam), which is not interesting for spammers. A better way to think of EmailJS in terms of security is not as a service that allows you to send emails from your code, but rather as a service that allows you to create a predefined set of emails via the dashboard, and then just trigger the emails from the code. This is quite similar to how emails are usually sent via a proprietary server code and how products like Intercom or customer.io are working.

* Works but requires some modifications